Added a bunch of regular expressions
Is this enough to justify a commit? Probably not. Am I going to commit anyway to prevent an "everything" commit? Absolutely. Can't actually test anything because I'm on a work machine so I'll do that later.

_________________________________

INT. PROMENADE

SISKO walks briskly through the premenade as an AMBASSADOR WITH A BAD HAIRCUT AND POINTED EARS follows close behind, attempted to have a conversation.

AMBASSADOR
Commander, I need to know the status of Ambassador J'Tel, her Mission is -

SISKO
-Of utmost important, yes. You've informed me. Unfortunately, we can't send a vessel to rendezvous with her while the energy fluctuations persist. It's too dangerous.