Fixed some issues with lexer, finished parse logic
Still no script for now, but stuff is working! I wasn't handling boolean equal properly at all and I ran into a recursion issue for... an embarassing amount of time because of that.

This all desparately needs polishing but I need to look at anything but a CST for a little while. (Especially before I do more testing and realize that only my few examples worked....)