Name Checking and type checking functional
 In crunch so we're nixing the script again but we'll get there.

I have to go back through this and add logs, but for now I'm just pleased that this isn't exploding in my face. It isn't done by any means.

We don't create a scope table yet and we don't have a way to go back and give warnings, BUT we're getting there!