Print, Variable Declaration, Assignment should all work
Now I have to do if and while and fight a jump table so..... we'll see what happens. I don't even have that sense of programming dread anymore.

I think my brain has finally realized that this is the last thing between me and graduation and I've just dissociated entirely.